optional params
in our team, we had a few times the problem, that
the developer want to execute a specific query,
that is not using any parameter.

now it is necessary to pass an empty object or array
to the methods. (mostly get list) but they will not
receive a Bearcat object, but directly the mySQL result.

it will also course that the used mySQL-connection
will not get released to the pool. (is locket forever)